iter14

+++ internal/repository/pqsqlrepo.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"github.com/Alheor/shorturl/internal/config"
 	"github.com/Alheor/shorturl/internal/randomname"
+	"github.com/Alheor/shorturl/internal/userauth"
 	"github.com/jackc/pgerrcode"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
@@ -38,11 +39,11 @@ func (pg *Postgres) Init(ctx context.Context) error {
 	return nil
 }
 
-func (pg *Postgres) Add(ctx context.Context, id string, value string) error {
+func (pg *Postgres) Add(ctx context.Context, user *userauth.User, id string, value string) error {
 
 	_, err := pg.Conn.Exec(ctx,
-		"INSERT INTO "+tableName+" (short_key, original_url) VALUES (@shortKey, @originalURL)",
-		pgx.NamedArgs{"shortKey": id, "originalURL": value},
+		"INSERT INTO "+tableName+" (user_id, short_key, original_url) VALUES (@userId, @shortKey, @originalURL)",
+		pgx.NamedArgs{"userId": user.ID, "shortKey": id, "originalURL": value},
 	)
 
 	if err == nil {
@@ -65,8 +66,8 @@ func (pg *Postgres) Add(ctx context.Context, id string, value string) error {
 
 		if uniqByOriginalURL {
 			row := pg.Conn.QueryRow(ctx,
-				"SELECT short_key FROM "+tableName+" WHERE original_url=@originalUrl",
-				pgx.NamedArgs{"originalUrl": value},
+				"SELECT short_key FROM "+tableName+" WHERE user_id = @userId AND original_url=@originalUrl",
+				pgx.NamedArgs{"userId": user.ID, "originalUrl": value},
 			)
 
 			var shortKey string
@@ -82,13 +83,13 @@ func (pg *Postgres) Add(ctx context.Context, id string, value string) error {
 	return err
 }
 
-func (pg *Postgres) Get(ctx context.Context, id string) (value string, error error) {
+func (pg *Postgres) Get(ctx context.Context, user *userauth.User, id string) (value string, error error) {
 
 	var originalURL string
 
 	row := pg.Conn.QueryRow(ctx,
-		"SELECT original_url FROM "+tableName+" WHERE short_key=@shortKey",
-		pgx.NamedArgs{"shortKey": id},
+		"SELECT original_url FROM "+tableName+" WHERE user_id = @userId AND short_key=@shortKey",
+		pgx.NamedArgs{"userId": user.ID, "shortKey": id},
 	)
 
 	err := row.Scan(&originalURL)
@@ -99,11 +100,11 @@ func (pg *Postgres) Get(ctx context.Context, id string) (value string, error err
 	return originalURL, nil
 }
 
-func (pg *Postgres) Remove(ctx context.Context, id string) {
+func (pg *Postgres) Remove(ctx context.Context, user *userauth.User, id string) {
 
 	_, err := pg.Conn.Exec(ctx,
-		"DELETE FROM "+tableName+" WHERE short_key=@shortKey",
-		pgx.NamedArgs{"shortKey": id},
+		"DELETE FROM "+tableName+" WHERE user_id = @userId AND short_key=@shortKey",
+		pgx.NamedArgs{"userId": user.ID, "shortKey": id},
 	)
 
 	if err != nil {
@@ -117,7 +118,7 @@ func (pg *Postgres) IsReady(ctx context.Context) bool {
 	return err == nil
 }
 
-func (pg *Postgres) AddBatch(ctx context.Context, in []BatchEl) error {
+func (pg *Postgres) AddBatch(ctx context.Context, user *userauth.User, in []BatchEl) error {
 
 	tx, err := pg.Conn.Begin(ctx)
 	if err != nil {
@@ -126,13 +127,13 @@ func (pg *Postgres) AddBatch(ctx context.Context, in []BatchEl) error {
 
 	var entries [][]any
 	for _, v := range in {
-		entries = append(entries, []any{v.ShortURL, v.OriginalURL})
+		entries = append(entries, []any{user.ID, v.ShortURL, v.OriginalURL})
 	}
 
 	_, err = pg.Conn.CopyFrom(
 		ctx,
 		pgx.Identifier{tableName},
-		[]string{"short_key", "original_url"},
+		[]string{"user_id", "short_key", "original_url"},
 		pgx.CopyFromRows(entries),
 	)
 
@@ -150,14 +151,50 @@ func (pg *Postgres) AddBatch(ctx context.Context, in []BatchEl) error {
 	return tx.Commit(ctx)
 }
 
+func (pg *Postgres) GetAll(ctx context.Context, user *userauth.User) (list []HistoryEl, error error) {
+
+	rows, err := pg.Conn.Query(ctx,
+		"SELECT short_key, original_url FROM "+tableName+" WHERE user_id = @userId",
+		pgx.NamedArgs{"userId": user.ID},
+	)
+
+	if err != nil {
+		return nil, errors.New(ErrNotFound)
+	}
+
+	defer rows.Close()
+
+	historyList := make([]HistoryEl, 0)
+	for rows.Next() {
+		var el HistoryEl
+		err = rows.Scan(&el.ShortURL, &el.OriginalURL)
+		if err != nil {
+			return nil, errors.New(ErrNotFound)
+		}
+
+		historyList = append(historyList, el)
+	}
+
+	err = rows.Err()
+	if err != nil {
+		return nil, errors.New(ErrNotFound)
+	}
+
+	return historyList, nil
+}
+
 func createDBSchema(ctx context.Context, conn *pgxpool.Pool) {
 
 	_, err := conn.Exec(ctx, `
 		CREATE TABLE IF NOT EXISTS `+tableName+` (
 		    id SERIAL NOT NULL PRIMARY KEY,
-		    short_key varchar(`+strconv.Itoa(randomname.ShortNameLength)+`) NOT NULL UNIQUE,
-		    original_url text NOT NULL UNIQUE
+		    user_id varchar(36) NOT NULL,
+		    short_key varchar(`+strconv.Itoa(randomname.ShortNameLength)+`) NOT NULL,
+		    original_url text NOT NULL
 		);
+
+		CREATE UNIQUE INDEX IF NOT EXISTS `+tableName+`_user_id_short_key_unique_idx ON `+tableName+` (user_id, short_key);
+		CREATE UNIQUE INDEX IF NOT EXISTS `+tableName+`_user_id_original_url_unique_idx ON `+tableName+` (user_id, original_url);
 	`)
 
 	if err != nil {
